LPS-199304 Extract endpoint configuration fields to components
