+++ modules/apps/headless/headless-builder/headless-builder-web/src/main/resources/META-INF/resources/js/components/EditEndpointConfiguration.tsx
@@ -3,13 +3,11 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {Text} from '@clayui/core';
 import ClayForm from '@clayui/form';
-import ClayIcon from '@clayui/icon';
-import {ClayTooltipProvider} from '@clayui/tooltip';
-import {sub} from 'frontend-js-web';
 import React, {Dispatch, SetStateAction, useEffect, useState} from 'react';
 
+import FiltersAndSorting from './endpointComponents/FiltersAndSorting';
+import PathParameterConfiguration from './endpointComponents/PathParameterConfiguration';
 import {Select} from './fieldComponents/Select';
 import {getAllItems} from './utils/fetchUtil';
 
@@ -79,14 +77,6 @@ export default function EditEndpointConfiguration({
 		);
 	};
 
-	const endpointFiltersInstruction = Liferay.Language.get(
-		'add-a-filter-using-odata'
-	);
-
-	const endpointSortInstruction = Liferay.Language.get(
-		'add-a-sort-using-odata'
-	);
-
 	return (
 		<ClayForm>
 			<ClayForm.Group>
@@ -107,121 +97,15 @@ export default function EditEndpointConfiguration({
 				/>
 			</ClayForm.Group>
 
-			<ClayForm.Group>
-				<label htmlFor="endpointFiltersField">
-					{Liferay.Language.get('filters')}
-
-					<ClayTooltipProvider>
-						<span
-							data-tooltip-align="top"
-							title={`${Liferay.Language.get(
-								'odata-cannot-exceed-1000-characters'
-							)} ${sub(
-								Liferay.Language.get(
-									'remember-not-to-include-x'
-								),
-								'?filter='
-							)}`}
-						>
-							&nbsp;
-							<ClayIcon symbol="question-circle-full" />
-						</span>
-					</ClayTooltipProvider>
-				</label>
-
-				<Text as="p" id="hostTextPreview" size={2} weight="lighter">
-					/?filter=
-				</Text>
-
-				<textarea
-					aria-label={endpointFiltersInstruction}
-					autoComplete="off"
-					className="form-control"
-					id="endpointFiltersField"
-					onChange={({target: {value}}) =>
-						setData((previousData) => ({
-							...previousData,
-							...(value !== ''
-								? {
-										apiEndpointToAPIFilters: [
-											{
-												...(previousData
-													.apiEndpointToAPIFilters?.[0]
-													?.id && {
-													id:
-														previousData
-															.apiEndpointToAPIFilters?.[0]
-															.id,
-												}),
-												oDataFilter: value,
-											},
-										],
-								  }
-								: {apiEndpointToAPIFilters: []}),
-						}))
-					}
-					placeholder={endpointFiltersInstruction}
-					value={data.apiEndpointToAPIFilters?.[0]?.oDataFilter}
+			{data.retrieveType?.key === 'singleElement' ? (
+				<PathParameterConfiguration
+					data={data}
+					selectedResponseBodySchema={selectedResponseBodySchema}
+					setData={setData}
 				/>
-			</ClayForm.Group>
-
-			<ClayForm.Group>
-				<label htmlFor="endpointSortingField">
-					{Liferay.Language.get('sorting')}
-
-					<ClayTooltipProvider>
-						<span
-							data-tooltip-align="top"
-							title={`${Liferay.Language.get(
-								'odata-cannot-exceed-1000-characters'
-							)} ${sub(
-								Liferay.Language.get(
-									'remember-not-to-include-x'
-								),
-								'?sort='
-							)}`}
-						>
-							&nbsp;
-							<ClayIcon symbol="question-circle-full" />
-						</span>
-					</ClayTooltipProvider>
-				</label>
-
-				<Text as="p" id="hostTextPreview" size={2} weight="lighter">
-					/?sort=
-				</Text>
-
-				<textarea
-					aria-label={endpointSortInstruction}
-					autoComplete="off"
-					className="form-control"
-					id="endpointSortingField"
-					onChange={({target: {value}}) =>
-						setData((previousData) => ({
-							...previousData,
-							...(value !== ''
-								? {
-										apiEndpointToAPISorts: [
-											{
-												...(previousData
-													.apiEndpointToAPISorts?.[0]
-													?.id && {
-													id:
-														previousData
-															.apiEndpointToAPISorts?.[0]
-															.id,
-												}),
-												oDataSort: value,
-											},
-										],
-								  }
-								: {apiEndpointToAPISorts: []}),
-						}))
-					}
-					placeholder={endpointSortInstruction}
-					value={data.apiEndpointToAPISorts?.[0]?.oDataSort}
-				/>
-			</ClayForm.Group>
+			) : (
+				<FiltersAndSorting data={data} setData={setData} />
+			)}
 		</ClayForm>
 	);
 }
